Rename MainController to BaseController
